forego aspect ratio and enter image size on thumbnail creation
sometimes the image is created before the aspect ratio is calculated and new size set, which results in a blank image.  On future loads the image is fine.  Setting image size on object instantiation removes this problem at the cost of not maintaining aspect ratio.  Alternative seems to be set image size on each object after they are created, which isn't easy :(